game implemented

+++ src/engine/actor.h
@@ -11,7 +11,7 @@ namespace engine {
 
         virtual void draw(const render &r) = 0;
 
-        virtual void update(double delta_time) = 0;
+        virtual void update(float delta_time) = 0;
 
         virtual std::unique_ptr<actor> clone();
